+++ Tests/Editor/ObjectMenuCreationTests.cs
@@ -1,7 +1,4 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 using UnityEngine.TestTools;
 using UnityEditor;
 
@@ -13,7 +10,11 @@ namespace UnityEngine.U2D
         public void ExecuteMenuCommandCreatesGameObjectWithPixelPerfectCamera()
         {
             var transformCount = Object.FindObjectsOfType<Transform>();
+#if ENABLE_URP_14_0_0_OR_NEWER
+            EditorApplication.ExecuteMenuItem("GameObject/2D Object/Pixel Perfect Camera (URP)");
+#else
             EditorApplication.ExecuteMenuItem("GameObject/2D Object/Pixel Perfect Camera");
+#endif
             LogAssert.NoUnexpectedReceived();
             Assert.True(Object.FindObjectsOfType<Transform>().Length > transformCount.Length);
         }